Make OCI image assembly work on every platform
Split embarcadero.oci's filesystem image-opening (`ImageOpenable`, which
reaches disk through bitumen's `TarOpenable`) into a JVM-only source set,
leaving in-memory image assembly and reading — building tar layers and
hashing blobs — cross-platform. A `jsSources` override excludes the JVM
source set from the Scala.js build, and `jsDeps` drops the non-JS
`bitumen.jvm` dependency, so OCI images can now be built and read in the
browser or a Wasm guest. Opening an image from a filesystem path still
requires the JVM.
