+++ build.mill
@@ -1070,13 +1070,15 @@ object diuretic extends Library:
       settings.cc(super.scalacOptions())
 
 object embarcadero extends Library:
-  // Building and hashing OCI image contents is cross-platform now that `bitumen.core` and
-  // `gastronomy` are, but opening an image from disk (`ImageHandle`) still needs `bitumen.jvm`, so
-  // the module stays JVM-only until it too is split into an in-memory core and a filesystem backend.
+  // Assembling and reading OCI images in memory — building tar layers and hashing blobs — is
+  // cross-platform, so an image can be built in the browser or a Wasm guest. Only opening an image
+  // from a filesystem *path* needs `bitumen.jvm`; that lives in the JVM-only `oci-jvm` source set,
+  // which `jsSources` excludes (and `jsDeps` drops the non-JS `bitumen.jvm` dependency for JS).
   object oci
       extends Component(bitumen.core, bitumen.jvm, jacinta.core, gastronomy.core, monotonous.core,
         gesticulate.core, aperture.core, pneumatic.flate):
-    override def scalaJs = false
+    override def sources = Task.Sources(moduleDir, moduleDir / os.up / "oci-jvm")
+    override def jsSources = Task.Sources(moduleDir)
     override def scalacOptions = Task:
       settings.cc(settings.maxInlines(super.scalacOptions() ++ Seq("-Yno-predef", "-Yimports:java.lang,scala,proscenium"), "64"))
   object containerd extends Component(obligatory.grpc, cordillera.core, coaxial.core, locomotion.core, anticipation.time, aperture.core):
