+++ lib/embarcadero/src/oci-jvm/embarcadero.ImageOpenable.scala
@@ -30,7 +30,40 @@
 ┃                                                                                                  ┃
 ┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┛
                                                                                                   */
-package soundness
+package embarcadero
 
-export embarcadero.{ContainerConfig, Descriptor, History, Image, ImageConfig, ImageDataOpenable,
-    ImageHandle, Index, Layer, Oci, OciError, RootFs}
+import anticipation.*
+import aperture.*
+import bitumen.*
+import contingency.*
+import prepositional.*
+import turbulence.*
+import zephyrine.*
+
+// Opening a filesystem *path* as an OCI image, delegating the TAR bracket to bitumen's disk-backed
+// `TarOpenable`. Split from `embarcadero.oci`'s cross-platform sources because it needs
+// `bitumen.jvm`; the in-memory `data.open[Image]` (via `ImageDataOpenable`) stays in the core.
+class ImageOpenable[path: Abstractable across Paths to Text]
+  ( using Tactic[OciError], Tactic[TarError], Tactic[StreamError] )
+extends Openable:
+
+  type Self = path
+  type Form = Image
+  type Operand = Nothing
+  type Result = ImageHandle
+
+  def open[grants <: Grant, result]
+    ( value: path, mode: Mode granting grants, flags: List[Nothing] )
+    ( block: ((ImageHandle & Granting[grants])^) ?=> result )
+  :   result =
+
+    if mode.atoms.contains(Write) then abort(OciError(OciError.Reason.WriteUnsupported))
+
+    TarOpenable[path]().open(value, mode, Nil): tar ?=>
+      block(using new ImageHandle(tar.entries) with Granting[grants] {})
+
+// Re-exported through `soundness.*`, so `path.open[Image]` resolves on the JVM as before.
+given openable: [path: Abstractable across Paths to Text]
+=>  ( Tactic[OciError], Tactic[TarError], Tactic[StreamError] )
+=>  ( ImageOpenable[path]^ ) =
+  ImageOpenable[path]
