+++ lib/embarcadero/src/oci-jvm/soundness_embarcadero_oci_jvm.scala
@@ -32,5 +32,4 @@
                                                                                                   */
 package soundness
 
-export embarcadero.{ContainerConfig, Descriptor, History, Image, ImageConfig, ImageDataOpenable,
-    ImageHandle, Index, Layer, Oci, OciError, RootFs}
+export embarcadero.{ImageOpenable, openable}
